+++ app/src/main/java/com/byeboo/app/domain/model/QuestDetailModel.kt
@@ -2,8 +2,8 @@ package com.byeboo.app.domain.model
 
 data class QuestDetailModel(
     val step: String,
-    val stepNumber: Int,
-    val questNumber: Int,
+    val stepNumber: Long,
+    val questNumber: Long,
     val questStyle: QuestStyle,
     val question: String
 )
